+++ src/cobib/importers/__init__.py
@@ -7,3 +7,4 @@ The abstract interface which should be implemented is defined in `cobib.importer
 """
 
 from .bibtex import BibtexImporter as BibtexImporter
+from .yaml import YAMLImporter as YAMLImporter